+++ blocklets/ai-studio/src/components/application/ApplicationHeader.tsx
@@ -4,10 +4,19 @@ import { HeaderWidgetCreator, useHeaderState } from '@blocklet/pages-kit/builtin
 import Header from '@blocklet/ui-react/lib/Header';
 import { Avatar, Box, Tooltip } from '@mui/material';
 import { useEffect, useMemo } from 'react';
+import { joinURL } from 'ufo';
 
+import { AIGNE_STUDIO_MOUNT_POINT } from '../../libs/constants';
+import { Deployment } from '../../libs/deployment';
 import { PlanUpgradeButton } from '../multi-tenant-restriction';
 
-export default function ApplicationHeader({ application, meta }: { application?: Agent; meta?: any }) {
+export default function ApplicationHeader({
+  application,
+  meta,
+}: {
+  application?: Agent & { deployment: Deployment };
+  meta?: any;
+}) {
   const { addons, add: addHeader, delete: deleteHeader } = useHeaderState();
 
   useEffect(() => {
@@ -30,6 +39,9 @@ export default function ApplicationHeader({ application, meta }: { application?:
         <Avatar variant="rounded" src={profile.icon} sx={{ width: 'auto', height: '100%' }} />
       </Tooltip>
     );
+  }
+  if (application?.deployment?.id) {
+    props.homeLink = joinURL(AIGNE_STUDIO_MOUNT_POINT, 'apps', application.deployment.id);
   }
 
   return (